DEV-1983 Add multistage-build Dockerfile
Build a Docker image containing the statically-linked Rust binary. A
muslrust image is used as the builder image as it compiles openssl and
links it against musl, which is needed to create the statically-linked
binary. The application image is based on a scratch image containing just
that binary.

The server was bound to 127.0.0.1 (localhost). This means that requests
can only come from inside the container. Binding to 0.0.0.0 ensures that
the requests are allowed from any IP address.