DEV-2588 Allow PREMIS v3 diff events
The PREMIS v3 (diff) events have a different format. The `premis:eventDetail`
node is encapsulated in a `premis:eventDetailInformation` node. As a quick
fix make `premis:eventDetail` optional in the deserialization struct.

Do note that the struct is not used for XML serialization. The data payload
of the cloudevent is the literal XML string of the incoming webhook event.